requirements.txt
fastapi
uvicorn
selenium
webdriver-manager
python-dotenv
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pillow
gspread
oauth2client